spectators: Move build_login_link out of hash_util.ts.
