+++ web/tests/hash_util.test.js
@@ -8,6 +8,7 @@ const {run_test} = require("./lib/test");
 const hash_util = zrequire("hash_util");
 const stream_data = zrequire("stream_data");
 const people = zrequire("people");
+const spectators = zrequire("spectators");
 
 const hamlet = {
     user_id: 15,
@@ -191,5 +192,5 @@ run_test("test_search_public_streams_notice_url", () => {
 
 run_test("test_current_hash_as_next", () => {
     window.location.hash = "#foo";
-    assert.equal(hash_util.current_hash_as_next(), "next=/%23foo");
+    assert.equal(spectators.current_hash_as_next(), "next=/%23foo");
 });
